+++ core/execution/amm/pool.go
@@ -232,35 +232,39 @@ func (p *Pool) setCurves(
 	linearSlippage num.Decimal,
 ) {
 	// convert the bounds into asset precision
-	lowerBound := num.UintZero().Mul(p.Parameters.LowerBound, p.priceFactor)
 	base := num.UintZero().Mul(p.Parameters.Base, p.priceFactor)
-	upperBound := num.UintZero().Mul(p.Parameters.UpperBound, p.priceFactor)
 
-	p.lower = generateCurve(
-		p.sqrt,
-		p.Commitment.Clone(),
-		lowerBound,
-		base,
-		lowerBound,
-		rfs.Long,
-		sfs.InitialMargin,
-		linearSlippage,
-		p.Parameters.MarginRatioAtLowerBound,
-		p.positionFactor,
-	)
+	if p.Parameters.LowerBound != nil {
+		lowerBound := num.UintZero().Mul(p.Parameters.LowerBound, p.priceFactor)
+		p.lower = generateCurve(
+			p.sqrt,
+			p.Commitment.Clone(),
+			lowerBound,
+			base,
+			lowerBound,
+			rfs.Long,
+			sfs.InitialMargin,
+			linearSlippage,
+			p.Parameters.MarginRatioAtLowerBound,
+			p.positionFactor,
+		)
+	}
 
-	p.upper = generateCurve(
-		p.sqrt,
-		p.Commitment.Clone(),
-		base.Clone(),
-		upperBound,
-		upperBound,
-		rfs.Short,
-		sfs.InitialMargin,
-		linearSlippage,
-		p.Parameters.MarginRatioAtUpperBound,
-		p.positionFactor,
-	)
+	if p.Parameters.UpperBound != nil {
+		upperBound := num.UintZero().Mul(p.Parameters.UpperBound, p.priceFactor)
+		p.upper = generateCurve(
+			p.sqrt,
+			p.Commitment.Clone(),
+			base.Clone(),
+			upperBound,
+			upperBound,
+			rfs.Short,
+			sfs.InitialMargin,
+			linearSlippage,
+			p.Parameters.MarginRatioAtUpperBound,
+			p.positionFactor,
+		)
+	}
 }
 
 // impliedPosition returns the position of the pool if its fair-price were the given price. `l` is
@@ -284,10 +288,10 @@ func (p *Pool) OrderbookShape(from, to *num.Uint) ([]*types.Order, []*types.Orde
 	buys := []*types.Order{}
 	sells := []*types.Order{}
 
-	if from == nil {
+	if from == nil && p.lower != nil {
 		from = p.lower.low
 	}
-	if to == nil {
+	if to == nil && p.upper != nil {
 		to = p.upper.high
 	}
 
@@ -296,6 +300,9 @@ func (p *Pool) OrderbookShape(from, to *num.Uint) ([]*types.Order, []*types.Orde
 	fairPrice := p.fairPrice()
 
 	ordersFromCurve := func(cu *curve, from, to *num.Uint) {
+		if cu == nil {
+			return
+		}
 		from = num.Max(from, cu.low)
 		to = num.Min(to, cu.high)
 		price := from
@@ -340,18 +347,36 @@ func (p *Pool) VolumeBetweenPrices(side types.Side, price1 *num.Uint, price2 *nu
 	// get the curve based on the pool's current position, if the position is zero we take the curve the trade will put us in
 	// e.g trading with an incoming buy order will make the pool short, so we take the upper curve.
 	var cu *curve
-	if pos < 0 || (pos == 0 && side == types.SideBuy) {
-		cu = p.upper
-	} else {
+	var high, low *num.Uint
+	if p.upper != nil && p.lower != nil {
+		if pos < 0 || (pos == 0 && side == types.SideBuy) {
+			cu = p.upper
+		} else {
+			cu = p.lower
+		}
+		high, low = cu.high, cu.low
+	} else if p.upper == nil {
 		cu = p.lower
+		high, low = cu.high, cu.low
+		// we should treat this as an upper curve
+		if pos < 0 || (pos == 0 && side == types.SideBuy) {
+			high, low = low, high
+		}
+	} else {
+		cu = p.upper
+		high, low = cu.high, cu.low
+		// we should treat this as an lower curve
+		if pos > 0 || (pos == 0 && side != types.SideBuy) {
+			high, low = low, high
+		}
 	}
 
 	if price1 == nil {
-		st = cu.low
+		st = low
 	}
 
 	if price2 == nil {
-		nd = cu.high
+		nd = high
 	}
 
 	if st.EQ(nd) {
@@ -363,16 +388,16 @@ func (p *Pool) VolumeBetweenPrices(side types.Side, price1 *num.Uint, price2 *nu
 	}
 
 	// there is no volume outside of the bounds for the curve so we snap st, nd to the boundaries
-	st = num.Max(st, cu.low)
-	nd = num.Min(nd, cu.high)
+	st = num.Max(st, low)
+	nd = num.Min(nd, high)
 	if st.GTE(nd) {
 		return 0
 	}
 
 	// abs(P(st) - P(nd))
 	volume, _ := num.UintZero().Delta(
-		impliedPosition(p.sqrt(st), p.sqrt(cu.high), cu.l),
-		impliedPosition(p.sqrt(nd), p.sqrt(cu.high), cu.l),
+		impliedPosition(p.sqrt(st), p.sqrt(high), cu.l),
+		impliedPosition(p.sqrt(nd), p.sqrt(high), cu.l),
 	)
 
 	if p.closing() {
@@ -449,7 +474,16 @@ func (p *Pool) getPosition() (int64, *num.Uint) {
 // x = P + (cc * rf) / sqrt(pl) + L / sqrt(pl),
 // y = abs(P) * average-entry + L * sqrt(pl).
 func (p *Pool) virtualBalancesShort(pos int64, ae *num.Uint) (num.Decimal, num.Decimal) {
+	var high, low *num.Uint
 	cu := p.upper
+	if p.upper != nil {
+		high = p.upper.high
+		low = p.upper.low
+	} else {
+		cu = p.lower
+		high = p.lower.low
+		low = p.lower.high
+	}
 	balance := p.getBalance()
 
 	// lets start with x
@@ -458,10 +492,10 @@ func (p *Pool) virtualBalancesShort(pos int64, ae *num.Uint) (num.Decimal, num.D
 	term1x := num.DecimalFromInt64(-pos)
 
 	// cc * rf / pu
-	term2x := cu.rf.Mul(num.DecimalFromUint(balance)).Div(num.DecimalFromUint(cu.high))
+	term2x := cu.rf.Mul(num.DecimalFromUint(balance)).Div(num.DecimalFromUint(high))
 
 	// L / sqrt(pl)
-	term3x := cu.l.ToDecimal().Div(p.sqrt(cu.high))
+	term3x := cu.l.ToDecimal().Div(p.sqrt(high))
 
 	// x = P + (cc * rf / pu) + (L / sqrt(pl))
 	x := term2x.Add(term3x).Sub(term1x)
@@ -472,7 +506,7 @@ func (p *Pool) virtualBalancesShort(pos int64, ae *num.Uint) (num.Decimal, num.D
 	term1y := ae.Mul(ae, num.NewUint(uint64(-pos)))
 
 	// L * sqrt(pl)
-	term2y := cu.l.ToDecimal().Mul(p.sqrt(cu.low))
+	term2y := cu.l.ToDecimal().Mul(p.sqrt(low))
 
 	// y = abs(P) * average-entry + L * pl
 	y := term1y.ToDecimal().Add(term2y)
@@ -484,20 +518,27 @@ func (p *Pool) virtualBalancesShort(pos int64, ae *num.Uint) (num.Decimal, num.D
 // x = P + (L / sqrt(pu)),
 // y = L * (sqrt(pu) - sqrt(pl)) - P * average-entry + (L * sqrt(pl)).
 func (p *Pool) virtualBalancesLong(pos int64, ae *num.Uint) (num.Decimal, num.Decimal) {
+	var high *num.Uint
 	cu := p.lower
+	if p.lower != nil {
+		high = cu.high
+	} else {
+		cu = p.upper
+		high = cu.low
+	}
 	// lets start with x
 
 	// P
 	term1x := num.DecimalFromInt64(pos)
 
 	// L / sqrt(pu)
-	term2x := cu.l.ToDecimal().Div(p.sqrt(cu.high))
+	term2x := cu.l.ToDecimal().Div(p.sqrt(high))
 	x := term1x.Add(term2x)
 
 	// now lets move to y
 
 	// L * (sqrt(pu) - sqrt(pl)) + (L * sqrt(pl)) => L * sqrt(pu)
-	term1y := cu.l.ToDecimal().Mul(p.sqrt(cu.high))
+	term1y := cu.l.ToDecimal().Mul(p.sqrt(high))
 
 	// P * average-entry
 	term2y := ae.Mul(ae, num.NewUint(uint64(pos)))
@@ -510,7 +551,10 @@ func (p *Pool) virtualBalancesLong(pos int64, ae *num.Uint) (num.Decimal, num.De
 func (p *Pool) fairPrice() *num.Uint {
 	pos, ae := p.getPosition()
 	if pos == 0 {
-		return p.lower.high.Clone()
+		if p.lower != nil {
+			return p.lower.high.Clone()
+		}
+		return p.upper.low.Clone()
 	}
 
 	x, y := p.virtualBalances(pos, ae, types.SideUnspecified)