+++ core/integration/setup_test.go
@@ -522,5 +522,9 @@ func (e *executionTestSetup) registerNetParamsCallbacks() error {
 			Param:   netparams.MarketLiquidityEquityLikeShareFeeFraction,
 			Watcher: execsetup.executionEngine.OnMarketLiquidityEquityLikeShareFeeFractionUpdate,
 		},
+		netparams.WatchParam{
+			Param:   netparams.MarketAMMMinCommitmentQuantum,
+			Watcher: execsetup.executionEngine.OnMarketAMMMinCommitmentQuantum,
+		},
 	)
 }