+++ core/integration/steps/the_following_trades_happened.go
@@ -27,6 +27,7 @@ import (
 )
 
 func TheFollowingTradesShouldBeExecuted(
+	exec Execution,
 	broker *stubs.BrokerStub,
 	table *godog.Table,
 ) error {
@@ -34,13 +35,24 @@ func TheFollowingTradesShouldBeExecuted(
 
 	for _, row := range parseExecutedTradesTable(table) {
 		buyer := row.MustStr("buyer")
+		buyerIn := buyer
 		seller := row.MustStr("seller")
+		sellerIn := seller
 		price := row.MustU64("price")
 		size := row.MustU64("size")
 		aggressorRaw := row.Str("aggressor side")
 		aggressor, aerr := Side(aggressorRaw)
 		if aggressorRaw != "" && aerr != nil {
 			return aerr
+		}
+		// remap buyer/seller to AMM IDs
+		if row.HasColumn("is amm") && row.MustBool("is amm") {
+			if id, ok := exec.GetAMMSubAccountID(buyer); ok {
+				buyer = id
+			}
+			if id, ok := exec.GetAMMSubAccountID(seller); ok {
+				seller = id
+			}
 		}
 
 		buyerFee, hasBuyerFee := row.U64B("buyer fee")
@@ -97,7 +109,7 @@ func TheFollowingTradesShouldBeExecuted(
 			}
 		}
 		if !found {
-			return errMissingTrade(buyer, seller, price, size)
+			return errMissingTrade(buyerIn, sellerIn, price, size)
 		}
 	}
 	return err
@@ -140,6 +152,7 @@ func parseExecutedTradesTable(table *godog.Table) []RowWrapper {
 		"seller infrastructure fee referrer discount",
 		"seller liquidity fee referrer discount",
 		"seller maker fee referrer discount",
+		"is amm",
 	})
 }